l63: ELF loader code wired into the build (Lecture 63)
PeachOS64BitCourse commit d3ba5e8 ported with one SamOs
deviation on the target binary (see below).

Makefile / build.sh:
  Add build/loader/formats/elf.o + elfloader.o to FILES;
  build rules; mkdir build/loader/formats.

  mcopy now also drops blank.elf + shell.elf onto os.bin
  alongside SIMPLE.BIN.

src/loader/formats/elf.c:
  Add <stdint.h>, <stddef.h>.
  (void*)e_entry -> (void*)(uintptr_t)e_entry. ELF32's e_entry
  is 32 bits; we're on 64-bit void*.

src/loader/formats/elfloader.c:
  (int)header + header->e_shoff -> (uintptr_t)header +
    header->e_shoff (sheader, pheader). Same for p_vaddr casts
    in elf_process_phdr_pt_load. end_virtual_address widened
    to uintptr_t.

src/task/process.c:
  Re-include loader/formats/elfloader.h.

  process_load_elf un-stubbed: opens the file via elf_load,
  stashes the elf_file pointer on the process, sets filetype
  to ELF.

  process_map_elf un-stubbed: walks the ELF program headers,
  paging_map_to each PT_LOAD into the task's PML4 with U=1 (+W
  if PF_W). (uintptr_t) cast on p_vaddr for the same width
  reason.

src/kernel.c:
  SamOs DEVIATION: keep loading SIMPLE.BIN, not blank.elf.
  Two reasons:
    1. Our blank.elf is ELF64 (L62 toolchain change). The L63
       loader is still ELF32-shaped and rejects it via the
       EI_CLASS check. L65 refactors to ELF64.
    2. blank.elf is 2.1 MiB. Loading via ATA-PIO sector at a
       time takes well over 30 seconds in QEMU TCG. Not the
       hill to die on for the L63 milestone.
  L66 (after L65's refactor) is when we switch to loading the
  ELF for real.

Test: tests64/L63-elfloader-linked.sh asserts:
  - SIMPLE.BIN, BLANK.ELF, SHELL.ELF all visible via mdir
  - build/loader/formats/elf.o + elfloader.o present
  - kernel still boots to "user enter"

Verified on VGA:
  ...
  user enter

Doc: docs/64bit/chapters/L63-elf-loader-linked.md walks the
ELF32-vs-ELF64 mismatch, the cast widenings, and the SamOs
deviation on the load target.

Next: Lecture 64 - filesystem / disk bug fixes.

+++ src/loader/formats/elf.c
@@ -1,7 +1,12 @@
 #include "elf.h"
+#include <stdint.h>
+#include <stddef.h>
 
 void* elf_get_entry_ptr(struct elf_header* elf_header){
-    return (void*)elf_header->e_entry;
+    // L63 - widen the cast: e_entry is 32 bits in ELF32 headers,
+    // but void* is 64 bits on x86_64. uintptr_t intermediate
+    // suppresses GCC's int-to-pointer-cast-of-different-size.
+    return (void*)(uintptr_t)elf_header->e_entry;
 }
 
 uint32_t elf_get_entry(struct elf_header* elf_header){
